Added .toml file and modified dir setup
