+++ deeponet_acoustics/scripts/main3D_eval_accuracy.py
