+++ deeponet_acoustics/scripts/main3D_eval_speed.py
