+++ deeponet_acoustics/scripts/main3D_train.py
@@ -6,12 +6,14 @@
 #
 # Licensed under the MIT License.
 # ==============================================================================
-import json
+import argparse
+from train3D import train
 
-def parseSettings(path_to_json):
-    file_handle = open(path_to_json, "r")
-    data = file_handle.read()
-    json_obj = json.loads(data)
-    file_handle.close()
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--path_settings", type=str, required=True)
+    args = parser.parse_args()
+    train(args.path_settings)
 
-    return json_obj
+if __name__ == "__main__":
+    main()
