+++ deeponet_acoustics/utils/gaussian.py
